+++ bench/bench_helper.exs
@@ -1,2 +1,5 @@
-Code.require_file("scripts/insert_bench.exs", __DIR__)
-Code.require_file("scripts/all_bench.exs", __DIR__)
+# Micro benchmarks
+
+## Macro benchmarks need foundationdb and postgresql up and running
+Code.require_file("scripts/macro/insert_bench.exs", __DIR__)
+Code.require_file("scripts/macro/all_bench.exs", __DIR__)