+++ lib/ecto_foundationdb/layer/primary_kv_codec.ex
@@ -5,14 +5,23 @@ defmodule EctoFoundationDB.Layer.PrimaryKVCodec do
   alias EctoFoundationDB.Layer.Pack
   alias EctoFoundationDB.Options
   alias EctoFoundationDB.Tenant
+  alias EctoFoundationDB.Versionstamp
 
-  defstruct [:tuple]
+  defstruct [:tuple, :vs?, :packed]
 
   @metadata_key :multikey
 
-  def new(tuple) do
-    %__MODULE__{tuple: tuple}
+  def new(tuple, vs \\ false)
+
+  def new(tuple, vs) when is_tuple(tuple) do
+    %__MODULE__{tuple: tuple, vs?: vs}
   end
+
+  def new(key, vs) when is_binary(key) do
+    %__MODULE__{packed: key, vs?: vs}
+  end
+
+  def vs?(%__MODULE__{vs?: vs?}), do: vs?
 
   def stream_decode(kvs, tenant) do
     Stream.transform(
@@ -94,15 +103,45 @@ defmodule EctoFoundationDB.Layer.PrimaryKVCodec do
     end
   end
 
+  def with_packed_key(kv_codec = %{packed: packed}) when not is_nil(packed), do: kv_codec
+
+  def with_packed_key(kv_codec) do
+    %{kv_codec | packed: pack_key(kv_codec, nil)}
+  end
+
+  def with_unpacked_tuple(kv_codec = %{tuple: tuple}, _tenant) when not is_nil(tuple),
+    do: kv_codec
+
+  def with_unpacked_tuple(kv_codec, tenant) do
+    %{kv_codec | tuple: Tenant.unpack(tenant, kv_codec.packed)}
+  end
+
   def pack_key(kv_codec, t) do
-    kv_codec.tuple
-    |> add_codec_metadata(t)
-    |> :erlfdb_tuple.pack()
+    %__MODULE__{vs?: vs?} = kv_codec
+
+    tuple =
+      kv_codec.tuple
+      |> add_codec_metadata(t)
+
+    if vs?, do: :erlfdb_tuple.pack_vs(tuple), else: :erlfdb_tuple.pack(tuple)
+  end
+
+  def set_new_kvs(tx, %__MODULE__{vs?: true}, kvs) do
+    for {k, v} <- kvs do
+      :erlfdb.set_versionstamped_key(tx, k, v)
+    end
+  end
+
+  def set_new_kvs(tx, %__MODULE__{vs?: false}, kvs) do
+    for {k, v} <- kvs do
+      :erlfdb.set(tx, k, v)
+    end
   end
 
   def range(kv_codec) do
+    %__MODULE__{vs?: vs?} = kv_codec
     tuple = add_codec_metadata(kv_codec.tuple, nil)
-    start_key = :erlfdb_tuple.pack(tuple)
+    start_key = if vs?, do: :erlfdb_tuple.pack_vs(tuple), else: :erlfdb_tuple.pack(tuple)
     {start_key, start_key <> <<0xFF>>}
   end
 
@@ -136,7 +175,14 @@ defmodule EctoFoundationDB.Layer.PrimaryKVCodec do
         """
 
       :error ->
-        item = %DecodedKV{codec: Pack.primary_write_key_to_codec(tenant, k), data_object: v}
+        key_tuple = Tenant.unpack(tenant, k)
+        data_object = extract_complete_vs(key_tuple, v)
+
+        item = %DecodedKV{
+          codec: Pack.primary_write_key_to_codec(tenant, k),
+          data_object: data_object
+        }
+
         {[item], %{acc | key_tuple: nil, values: [], meta: nil}}
     end
   end
@@ -154,6 +200,7 @@ defmodule EctoFoundationDB.Layer.PrimaryKVCodec do
         case :erlang.crc32(fdb_value) do
           ^crc ->
             data_object = Pack.from_fdb_value(fdb_value)
+            data_object = extract_complete_vs(key_tuple, data_object)
 
             item =
               %DecodedKV{
@@ -177,6 +224,17 @@ defmodule EctoFoundationDB.Layer.PrimaryKVCodec do
         raise """
         Metadata error. Previous: #{inspect({n, i, crc})}, Encountered: #{inspect(other)}
         """
+    end
+  end
+
+  defp extract_complete_vs(key_tuple, data_object) do
+    [{pk_field, stored_pk} | data_object_rest] = data_object
+
+    # When incomplete versionstamp is stored on in the value, we need to retrieve the pk from the key
+    if Versionstamp.incomplete?(stored_pk) do
+      [{pk_field, Pack.get_vs_from_key_tuple(key_tuple)} | data_object_rest]
+    else
+      data_object
     end
   end
 end