+++ test/ecto/integration/fdb_api_counting_test.exs
@@ -5,11 +5,13 @@ defmodule Ecto.Integration.FdbApiCountingTest do
 
   alias EctoFoundationDB.Layer.MetadataVersion
   alias EctoFoundationDB.Tenant
+  alias EctoFoundationDB.Versionstamp
 
   alias Ecto.Adapters.FoundationDB
   alias Ecto.Integration.TestRepo
 
   alias EctoFoundationDB.ModuleToModuleTracer
+  alias EctoFoundationDB.Schemas.QueueItem
   alias EctoFoundationDB.Schemas.User
   alias EctoFoundationDB.Test.Util
 
@@ -25,6 +27,8 @@ defmodule Ecto.Integration.FdbApiCountingTest do
     {:erlfdb, :wait, 1},
     {:erlfdb, :wait_for_any, 1},
     {:erlfdb, :set, 3},
+    {:erlfdb, :set_versionstamped_key, 3},
+    {:erlfdb, :set_versionstamped_value, 3},
     {:erlfdb, :clear, 2},
     {:erlfdb, :clear_range, 3},
     {:erlfdb, :max, 3},
@@ -92,7 +96,7 @@ defmodule Ecto.Integration.FdbApiCountingTest do
              {EctoFoundationDB.Future, :wait_for_all_interleaving},
 
              # primary write
-             {EctoFoundationDB.Layer.TxInsert, :set},
+             {EctoFoundationDB.Layer.PrimaryKVCodec, :set},
 
              # index write
              {EctoFoundationDB.Indexer.Default, :set}
@@ -127,7 +131,7 @@ defmodule Ecto.Integration.FdbApiCountingTest do
              {EctoFoundationDB.Future, :wait_for_all_interleaving},
 
              # primary write, index write
-             {EctoFoundationDB.Layer.TxInsert, :set},
+             {EctoFoundationDB.Layer.PrimaryKVCodec, :set},
              {EctoFoundationDB.Indexer.Default, :set}
            ] == calls
   end
@@ -166,7 +170,7 @@ defmodule Ecto.Integration.FdbApiCountingTest do
              {EctoFoundationDB.Future, :wait_for_all_interleaving},
 
              # primary write, index write
-             {EctoFoundationDB.Layer.TxInsert, :set},
+             {EctoFoundationDB.Layer.PrimaryKVCodec, :set},
              {EctoFoundationDB.Indexer.Default, :set},
 
              # wait for app metadata version and claim_key
@@ -201,7 +205,7 @@ defmodule Ecto.Integration.FdbApiCountingTest do
              {EctoFoundationDB.Future, :wait_for_all_interleaving},
 
              # primary write, index write
-             {EctoFoundationDB.Layer.TxInsert, :set},
+             {EctoFoundationDB.Layer.PrimaryKVCodec, :set},
              {EctoFoundationDB.Indexer.Default, :set}
            ] == calls
   end
@@ -309,13 +313,13 @@ defmodule Ecto.Integration.FdbApiCountingTest do
 
     {calls, _} =
       with_erlfdb_calls(context.test, fn ->
-        TestRepo.transaction(
+        TestRepo.transactional(
+          tenant,
           fn ->
             f1 = TestRepo.async_get_by(User, name: "Alice")
             f2 = TestRepo.async_get_by(User, name: "Bob")
             TestRepo.await([f1, f2])
-          end,
-          prefix: tenant
+          end
         )
       end)
 
@@ -395,11 +399,11 @@ defmodule Ecto.Integration.FdbApiCountingTest do
              {EctoFoundationDB.Future, :wait_for_all_interleaving},
 
              # primary write
-             {EctoFoundationDB.Layer.TxInsert, :set},
+             {EctoFoundationDB.Layer.PrimaryKVCodec, :set},
 
              # multikey writes
-             {EctoFoundationDB.Layer.TxInsert, :set},
-             {EctoFoundationDB.Layer.TxInsert, :set},
+             {EctoFoundationDB.Layer.PrimaryKVCodec, :set},
+             {EctoFoundationDB.Layer.PrimaryKVCodec, :set},
 
              # index write
              {EctoFoundationDB.Indexer.Default, :set}
@@ -492,6 +496,38 @@ defmodule Ecto.Integration.FdbApiCountingTest do
 
              # clear :name index
              {EctoFoundationDB.Indexer.Default, :clear}
+           ] == calls
+  end
+
+  test "async_insert_all!", context do
+    tenant = context[:tenant]
+
+    # Prime metadata cache
+    TestRepo.transactional(tenant, fn tx ->
+      TestRepo.insert!(%QueueItem{id: Versionstamp.next(tx)})
+    end)
+
+    {calls, _eve} =
+      with_erlfdb_calls(context.test, fn ->
+        f =
+          TestRepo.transactional(tenant, fn ->
+            TestRepo.async_insert_all!(QueueItem, [%QueueItem{author: "test", data: "test"}])
+          end)
+
+        TestRepo.await(f)
+      end)
+
+    assert [
+             # we always get global metadataVersion
+             {EctoFoundationDB.Layer.MetadataVersion, :get},
+             {EctoFoundationDB.Future, :wait},
+
+             # set the primary write and the index
+             {EctoFoundationDB.Layer.PrimaryKVCodec, :set_versionstamped_key},
+             {EctoFoundationDB.Indexer.Default, :set_versionstamped_key},
+
+             # wait for versionstamp future
+             {EctoFoundationDB.Future, :wait_for_all_interleaving}
            ] == calls
   end
 end