+++ test/ecto/integration/max_value_size_test.exs
@@ -61,6 +61,7 @@ defmodule Ecto.Integration.MaxValueSizeTest do
 
       assert %User{} = TestRepo.get(User, user.id, prefix: tenant)
 
+      # @todo: verify user has the expected `:notes` data
       assert %User{} = TestRepo.get_by(User, [name: "Bob"], prefix: tenant)
 
       assert {:ok, _} = TestRepo.delete(user)